Neighbors and texture types finished

+++ src/map/automata.rs
@@ -2,7 +2,7 @@ use std::vec;
 
 use rand::Rng;
 
-pub fn generate(size: u16, density: f32) -> Vec<Vec<u8>> {
+pub fn generate(size: usize, density: f32) -> Vec<Vec<u8>> {
     let mut matr = vec::from_elem(vec::from_elem(0, size as usize), size as usize);
     
     let mut rng = rand::thread_rng();
